+++ dataflow/context.go
@@ -1,4 +1,4 @@
-package gout
+package dataflow
 
 import (
 	"github.com/guonaihong/gout/decode"
