+++ filter/init.go
@@ -1,0 +1,19 @@
+package gout
+
+import (
+	"github.com/guonaihong/gout/dataflow"
+	"math/rand"
+	"time"
+)
+
+var (
+	defaultBench = Bench{}
+	defaultRetry = Retry{}
+)
+
+func init() {
+	dataflow.Register("bench", &defaultBench)
+	dataflow.Register("retry", &defaultRetry)
+
+	rand.Seed(time.Now().UnixNano())
+}
