+++ import_rawtext.go
@@ -1,5 +1,9 @@
 package gout
 
+import (
+	"github.com/guonaihong/gout/dataflow"
+)
+
 type Text struct {
-	DataFlow
+	dataflow.DataFlow
 }
